fix(gitlab): move GitlabDeployment to  migration's archived package and add gitlab_id field
